Extract EventRepeatEnds as functional component
